requirements.txt
chainlit
websocket-client
pillow
python-dotenv
openai
requests
PyYAML
websockets
semantic-version